Change CrazyRandomSword hit functionality
The no parameter version of the Weapon class' hit() function has been
made virtual and a separate version of the function has been implemented
in CRSword after getting confirmation that the CRSword should randomize
hitPoints every single hit rather than on creation.

This is not handled in CRSword's constructor as that would only work if
the no parameter hit() was called once as the hitPoints would only be
randomized on creation, not if hit() was called repeatedly.